test(autobahn): fold rpc-only sidecar lifecycle into autobahn_test.go
The dedicated rpc_only_test.go file existed when there was a dedicated
RPCOnlyForwarding sub-test plus EVM-side helpers it owned. Since we
collapsed those into autobahn_test.go (every sub-test now goes through
the rpc-only sidecar), the file only carried the rpc-only sidecar
lifecycle — setupRPCOnlyNode, rpcOnlyRunning, rpcOnlyEVMReady, plus
the evmRPCInContainer helper they use for readiness.

That's symmetric with setupCluster / teardownCluster in
autobahn_test.go — both are "bring infrastructure up, wait for ready"
helpers called from TestMain. Move them next to teardownCluster (which
already takes responsibility for tearing down the rpc-only sidecar
alongside the validators) and delete rpc_only_test.go.

No behaviour change.
